pyproject.toml
[project]
name = "docktapus"
version = "0.1.0"
description = ""
authors = [
    {name = "mfowlerrr",email = "matty01fowler@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer (>=0.21.1,<0.22.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[tool.poetry]
packages = [{include = "docktapus", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
